+++ src/api/routes/config.rs
@@ -1,13 +1,13 @@
-use actix_web::{HttpResponse, post, Responder, ResponseError, web};
 use actix_web::web::Data;
+use actix_web::{post, web, HttpResponse, Responder, ResponseError};
 use serde::{Deserialize, Serialize};
 use tokio::sync::RwLock;
-use tracing_subscriber::EnvFilter;
 use tracing_subscriber::fmt::Formatter;
 use tracing_subscriber::reload::Handle;
+use tracing_subscriber::EnvFilter;
 
+use crate::api::http_error::ErrorResponse;
 use crate::ban_hammer::BanHammerDryRunner;
-use crate::http_error::ErrorResponse;
 
 #[derive(Debug, Serialize, Deserialize)]
 pub struct ConfigureRequest {
@@ -28,7 +28,11 @@ pub async fn configuration_handler(
     }
     if let Some(log_lvl) = q.0.log_level {
         if let Err(e) = h.modify(|e| *e = EnvFilter::new(log_lvl)) {
-            return Err(ErrorResponse { code: 400, reason: e.to_string(), details: None });
+            return Err(ErrorResponse {
+                code: 400,
+                reason: e.to_string(),
+                details: None,
+            });
         }
     }
