+++ src/api/routes/mod.rs
@@ -1,7 +1,7 @@
-pub use config::configuration_handler;
 pub use ban::process_ban;
+pub use config::configuration_handler;
 pub use unban::process_unban;
 
-mod config;
 mod ban;
+mod config;
 mod unban;
