+++ src/lib.rs
@@ -3,7 +3,6 @@ pub mod ban_hammer;
 pub mod config;
 pub mod dry_runner;
 pub mod errors;
-pub mod http_error;
 pub mod model;
 pub mod redis;
 pub mod telemetry;
